LayoutTests/fast/hidpi/image-srcset-data-srcset-expected.txt

PASS document.getElementById("foo").clientWidth==150 is true
This test passes if the image below looks like a 2x2 grid (yellow, gray, light green, dark green). In this case the srcset attribute has a base64 url, with the comma escaped.
